+++ packages/components/Iframe/index.ts
@@ -1,2 +1,3 @@
 export { Iframe } from './Iframe'
 export type { IframeProps } from './Iframe'
+export { IframeDemo } from './IframeDemo'