+++ src/pages/index/demo/components/iframe.tsx
@@ -1,0 +1,3 @@
+import { ZaIframeDemo } from '@zealous-admin/components/index'
+
+export default ZaIframeDemo